chore: refactor defaults
