+++ src/api/index.ts
@@ -1,4 +1,5 @@
 export * from './types.ts'
+export * from './defaults.ts'
 export * from './utils.ts'
 export * from './html.ts'
 export * from './maskable.ts'
