+++ src/utils.ts
@@ -1,4 +1,3 @@
-import type { PngOptions, ResizeOptions } from 'sharp'
 import type {
   Asset,
   AssetSize,
@@ -7,17 +6,7 @@ import type {
   ResolvedAssetSize,
   ResolvedAssets,
 } from './types.ts'
-
-export const defaultPngCompressionOptions: PngOptions = {
-  compressionLevel: 9,
-  quality: 60,
-}
-
-export const defaultPngOptions: Record<AssetType, { padding: number; resizeOptions: ResizeOptions }> = {
-  transparent: { padding: 0.05, resizeOptions: { fit: 'contain', background: 'transparent' } },
-  maskable: { padding: 0.3, resizeOptions: { fit: 'contain', background: 'white' } },
-  apple: { padding: 0.3, resizeOptions: { fit: 'contain', background: 'white' } },
-}
+import { defaultPngOptions } from './api/defaults.ts'
 
 export function toResolvedSize(size: AssetSize) {
   return {
